# docs/openapi.yaml
openapi: 3.0.0
info:
  title: 🚀 Edu Management API
  description: Education Management System API collection
  version: 1.0.0
servers:
  - url: http://{{base_url}}{{api_version}}
  - url: http://{{base_url}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
...
paths:
...
  /academic-years/:
    get:
      tags:
        - 📅 Academic Year
      summary: List Academic Years
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - 📅 Academic Year
      summary: Create Academic Year
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"name\": \"2024-25\",\n  \"start_date\":
                \"2024-06-01T00:00:00Z\",\n  \"end_date\":
                \"2025-03-31T00:00:00Z\"\n}"
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /academic-years/{id}:
    get:
      tags:
        - 📅 Academic Year
      summary: Get Academic Year By ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    put:
      tags:
        - 📅 Academic Year
      summary: Update Academic Year
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"name\": \"2024-25\",\n  \"start_date\":
                \"2024-06-01T00:00:00Z\",\n  \"end_date\":
                \"2025-04-30T00:00:00Z\"\n}"
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    delete:
      tags:
        - 📅 Academic Year
      summary: Delete Academic Year
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /academic-years/{id}/activate:
    patch:
      tags:
        - 📅 Academic Year
      summary: Set Academic Year Active
      requestBody:
        content: {}
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /attendance/:
    post:
      tags:
        - 🗓️ Student Attendance
      summary: Mark Single Attendance
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"student_enrollment_id\": \"{{enrollment_id}}\",\n 
                \"date\": \"2025-09-01T00:00:00Z\",\n  \"status\":
                \"PRESENT\",\n  \"remark\": null,\n  \"recorded_by_teacher_id\":
                \"{{teacher_id}}\"\n}"
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    get:
      tags:
        - 🗓️ Student Attendance
      summary: List Student Attendance
      parameters:
        - name: class_section_id
          in: query
          schema:
            type: string
          example: '{{class_section_id}}'
        - name: date_from
          in: query
          schema:
            type: string
          example: '2025-09-01'
        - name: date_to
          in: query
          schema:
            type: string
          example: '2025-09-30'
        - name: page
          in: query
          schema:
            type: integer
          example: '1'
        - name: page_size
          in: query
          schema:
            type: integer
          example: '20'
        - name: sort_by
          in: query
          schema:
            type: string
          example: date
        - name: sort_dir
          in: query
          schema:
            type: string
          example: desc
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /attendance/bulk:
    post:
      tags:
        - 🗓️ Student Attendance
      summary: Bulk Mark Attendance (Class)
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"class_section_id\": \"{{class_section_id}}\",\n 
                \"date\": \"2025-09-02T00:00:00Z\",\n 
                \"recorded_by_teacher_id\": \"{{teacher_id}}\",\n  \"records\":
                [\n    {\n      \"student_enrollment_id\":
                \"<enrollment_uuid_1>\",\n      \"status\": \"PRESENT\",\n     
                \"remark\": null\n    },\n    {\n     
                \"student_enrollment_id\": \"<enrollment_uuid_2>\",\n     
                \"status\": \"ABSENT\",\n      \"remark\": \"Sick leave\"\n   
                },\n    {\n      \"student_enrollment_id\":
                \"<enrollment_uuid_3>\",\n      \"status\": \"LATE\",\n     
                \"remark\": \"Arrived 20 mins late\"\n    }\n  ]\n}"
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /attendance/{attendance_id}:
    get:
      tags:
        - 🗓️ Student Attendance
      summary: Get Attendance by ID
      parameters:
        - name: attendance_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    put:
      tags:
        - 🗓️ Student Attendance
      summary: Update Attendance
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"status\": \"HALF_DAY\",\n  \"remark\": \"Left early due
                to medical appointment\"\n}"
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: attendance_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    delete:
      tags:
        - 🗓️ Student Attendance
      summary: Delete Attendance
      parameters:
        - name: attendance_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /attendance/class/{class_section_id}/summary:
    get:
      tags:
        - 🗓️ Student Attendance
      summary: Get Class Attendance Summary
      parameters:
        - name: date
          in: query
          schema:
            type: string
          example: '2025-09-01'
        - name: class_section_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...